% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{showDT}
\alias{showDT}
\title{Create an HTML table using DT package with fixed columns}
\usage{
showDT(data, ...)
}
\arguments{
\item{data}{data object (either a matrix or a data frame).}

\item{...}{Additional arguments used by dDT::atatable() function.}
}
\value{
returns an object of \code{datatables} and \code{htmlwidget}.
}
\description{
Create an HTML table using DT package with fixed columns
}
\examples{
showDT(iris)
}
